backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.3.0
beanie==1.29.0
cffi==1.17.1
click==8.2.0
cryptography==44.0.3
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-users==14.0.1
fastapi-users-db-beanie==4.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
lazy-model==0.2.0
makefun==1.16.0
motor==3.7.0
pwdlib==0.2.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.12.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
toml==0.10.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
